Rename `remove_unreachable` to `remove_unreachable_functions`
