+++ compiler/noirc_evaluator/src/ssa/opt/remove_unreachable_functions.rs
